pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.11",
    "alembic>=1.14.1",
    "anthropic>=0.45.2",
    "bcrypt>=4.2.1",
    "beautifulsoup4==4.13.3",
    "ccxt>=4.4.57",
    "discord-py>=2.4.0",
    "fastapi>=0.115.8",
    "jinja2>=3.1.5",
    "joblib==1.4.2",
    "keras>=2.14.0",
    "lightgbm==4.5.0",
    "newsapi-python>=0.2.7",
    "nltk>=3.9.1",
    "numpy>=2.2.2",
    "openai==1.61.1",
    "pandas-ta>=0.3.14b0",
    "pandas>=2.2.3",
    "plotly==6.0.0",
    "praw==7.8.1",
    "psycopg2-binary==2.9.10",
    "pyserum==0.5.0a0",
    "pytest>=8.3.4",
    "python-binance==1.0.19",
    "python-dotenv==1.0.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "requests==2.32.3",
    "scikit-learn==1.6.1",
    "scipy>=1.15.1",
    "solana==0.25.1",
    "solana-sdk==0.25.6",
    "sqlalchemy==2.0.38",
    "streamlit>=1.42.0",
    "telethon==1.38.1",
    "tensorflow==2.14.0",
    "torch>=2.6.0",
    "trafilatura>=2.0.0",
    "transformers>=4.48.3",
    "tweepy==4.15.0",
    "twilio==9.4.1",
    "uvicorn>=0.34.0",
    "websocket-client==1.7.0",
    "websockets>=10.4",
    "xgboost==2.1.4",
    "yfinance==0.2.52",
    "web3>=7.8.0",
    "python-telegram-bot>=20.2",
]

[[tool.uv.index]]
explicit = true
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
